fix: 사진 추가 모달창 상위 컴포넌트 Home -> EditSidebar로 이동
